// frontend/src/models/outwardRowModel.tsx
export interface OutwardRow {
    id: string;
    projectId: string;
    srNo: number;
    letterNo: string;
    letterDate: string;
    to: string;
    subject: string;
    attachmentDetails: string;
    actionTaken: string;
    storagePath: string;
    remarks: string;
    acknowledgement: string;
}